refactor(cli): split driver/resolution.rs by resolution phase
The 4109-line `crates/tsz-cli/src/driver/resolution.rs` was over double the
§19 hard 2000-line cap. Split into focused submodules grouped by resolution
phase. Pure relocation; no behavior change.

- `resolution/type_packages.rs` (474 lines) - type-reference + type-package
  resolution, implied resolution mode, default type roots.
- `resolution/discovery.rs` (1251 lines) - text-based + AST-based module
  specifier discovery, JSDoc imports, import attributes, import/export
  binding traversal.
- `resolution/path_resolution.rs` (668 lines) - module-specifier path
  resolution, path mapping, candidate expansion, normalization,
  duplicate-package redirects.
- `resolution/package_resolution.rs` (958 lines) - package.json /
  node_modules / SemVer / node module specifier / package entry resolution.
- `resolution/exports_imports.rs` (567 lines) - types-versions matching,
  exports/imports subpath resolution.

Root `resolution.rs` (299 lines) retains shared infrastructure
(`ModuleResolutionCache`, filesystem probe counters,
`package_relative_target_path`, misc helpers, the
`AmbientModuleDeclarationSpecifierPolicy` enum) and re-exports the public
API so external callers keep using `crate::driver::resolution::<item>`
unchanged.

All five submodules and the root are now under the 2000-line cap.

AgentName: opus47-vibrant-lovelace-jLVUr

Closes #10709
