+++ crates/tsz-cli/src/driver/resolution/exports_imports.rs
@@ -1,0 +1,567 @@
+use std::path::{Path, PathBuf};
+
+use crate::config::ResolvedCompilerOptions;
+use tsz::module_resolver::PackageType;
+
+#[allow(unused_imports)]
+use super::*;
+
+pub(crate) fn resolve_types_versions(
+    package_root: &Path,
+    subpath: &str,
+    types_versions: &serde_json::Value,
+    options: &ResolvedCompilerOptions,
+    package_type: Option<PackageType>,
+    resolution_cache: &mut ModuleResolutionCache,
+) -> Option<PathBuf> {
+    let compiler_version = types_versions_compiler_version(options);
+    let paths = select_types_versions_paths(types_versions, compiler_version)?;
+    let mut best_pattern: Option<&String> = None;
+    let mut best_value: Option<&serde_json::Value> = None;
+    let mut best_wildcard = String::new();
+    let mut best_specificity = 0usize;
+    let mut best_len = 0usize;
+
+    for (pattern, value) in paths {
+        let Some(wildcard) = match_types_versions_pattern(pattern, subpath) else {
+            continue;
+        };
+        let specificity = types_versions_specificity(pattern);
+        let pattern_len = pattern.len();
+        let is_better = match best_pattern {
+            None => true,
+            Some(current) => {
+                specificity > best_specificity
+                    || (specificity == best_specificity && pattern_len > best_len)
+                    || (specificity == best_specificity
+                        && pattern_len == best_len
+                        && pattern < current)
+            }
+        };
+
+        if is_better {
+            best_specificity = specificity;
+            best_len = pattern_len;
+            best_pattern = Some(pattern);
+            best_value = Some(value);
+            best_wildcard = wildcard;
+        }
+    }
+
+    let value = best_value?;
+
+    let mut targets = Vec::new();
+    match value {
+        serde_json::Value::String(value) => targets.push(value.as_str()),
+        serde_json::Value::Array(list) => {
+            for entry in list {
+                if let Some(value) = entry.as_str() {
+                    targets.push(value);
+                }
+            }
+        }
+        _ => {}
+    }
+
+    for target in targets {
+        let substituted = substitute_path_target(target, &best_wildcard);
+        if let Some(resolved) = resolve_package_entry(
+            package_root,
+            &substituted,
+            options,
+            package_type,
+            resolution_cache,
+        ) {
+            return Some(resolved);
+        }
+    }
+
+    None
+}
+
+pub(crate) fn select_types_versions_paths(
+    types_versions: &serde_json::Value,
+    compiler_version: SemVer,
+) -> Option<&serde_json::Map<String, serde_json::Value>> {
+    select_types_versions_paths_for_version(types_versions, compiler_version)
+}
+
+pub(crate) fn select_types_versions_paths_for_version(
+    types_versions: &serde_json::Value,
+    compiler_version: SemVer,
+) -> Option<&serde_json::Map<String, serde_json::Value>> {
+    let map = types_versions.as_object()?;
+    let mut best_score: Option<RangeScore> = None;
+    let mut best_key: Option<&str> = None;
+    let mut best_value: Option<&serde_json::Map<String, serde_json::Value>> = None;
+
+    for (key, value) in map {
+        let Some(value_map) = value.as_object() else {
+            continue;
+        };
+        let Some(score) = match_types_versions_range(key, compiler_version) else {
+            continue;
+        };
+        let is_better = match best_score {
+            None => true,
+            Some(best) => {
+                score > best
+                    || (score == best && best_key.is_none_or(|best_key| key.as_str() < best_key))
+            }
+        };
+
+        if is_better {
+            best_score = Some(score);
+            best_key = Some(key);
+            best_value = Some(value_map);
+        }
+    }
+
+    best_value
+}
+
+pub(crate) fn match_types_versions_pattern(pattern: &str, subpath: &str) -> Option<String> {
+    if !pattern.contains('*') {
+        return (pattern == subpath).then(String::new);
+    }
+
+    let star = pattern.find('*')?;
+    let (prefix, suffix) = pattern.split_at(star);
+    let suffix = &suffix[1..];
+
+    if !subpath.starts_with(prefix) || !subpath.ends_with(suffix) {
+        return None;
+    }
+
+    let start = prefix.len();
+    let end = subpath.len().saturating_sub(suffix.len());
+    if end < start {
+        return None;
+    }
+
+    Some(subpath[start..end].to_string())
+}
+
+pub(crate) fn types_versions_specificity(pattern: &str) -> usize {
+    if let Some(star) = pattern.find('*') {
+        star + (pattern.len() - star - 1)
+    } else {
+        pattern.len()
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd)]
+pub(crate) struct RangeScore {
+    pub(super) constraints: usize,
+    pub(super) min_version: SemVer,
+    pub(super) key_len: usize,
+}
+
+pub(crate) fn match_types_versions_range(
+    range: &str,
+    compiler_version: SemVer,
+) -> Option<RangeScore> {
+    let range = range.trim();
+    if range.is_empty() || range == "*" {
+        return Some(RangeScore {
+            constraints: 0,
+            min_version: SemVer::ZERO,
+            key_len: range.len(),
+        });
+    }
+
+    let mut best: Option<RangeScore> = None;
+    for segment in range.split("||") {
+        let segment = segment.trim();
+        let Some(score) =
+            match_types_versions_range_segment(segment, compiler_version, range.len())
+        else {
+            continue;
+        };
+        if best.is_none_or(|current| score > current) {
+            best = Some(score);
+        }
+    }
+
+    best
+}
+
+pub(crate) fn match_types_versions_range_segment(
+    segment: &str,
+    compiler_version: SemVer,
+    key_len: usize,
+) -> Option<RangeScore> {
+    if segment.is_empty() {
+        return None;
+    }
+    if segment == "*" {
+        return Some(RangeScore {
+            constraints: 0,
+            min_version: SemVer::ZERO,
+            key_len,
+        });
+    }
+
+    let mut min_version = SemVer::ZERO;
+    let mut constraints = 0usize;
+
+    for token in segment.split_whitespace() {
+        if token.is_empty() || token == "*" {
+            continue;
+        }
+        let (op, version) = parse_range_token(token)?;
+        if !compare_range(compiler_version, op, version) {
+            return None;
+        }
+        constraints += 1;
+        if matches!(op, RangeOp::Gt | RangeOp::Gte | RangeOp::Eq) && version > min_version {
+            min_version = version;
+        }
+    }
+
+    Some(RangeScore {
+        constraints,
+        min_version,
+        key_len,
+    })
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub(crate) enum RangeOp {
+    Gt,
+    Gte,
+    Lt,
+    Lte,
+    Eq,
+}
+
+pub(crate) fn parse_range_token(token: &str) -> Option<(RangeOp, SemVer)> {
+    let token = token.trim();
+    if token.is_empty() {
+        return None;
+    }
+
+    let (op, rest) = if let Some(rest) = token.strip_prefix(">=") {
+        (RangeOp::Gte, rest)
+    } else if let Some(rest) = token.strip_prefix("<=") {
+        (RangeOp::Lte, rest)
+    } else if let Some(rest) = token.strip_prefix('>') {
+        (RangeOp::Gt, rest)
+    } else if let Some(rest) = token.strip_prefix('<') {
+        (RangeOp::Lt, rest)
+    } else if let Some(rest) = token.strip_prefix('=') {
+        (RangeOp::Eq, rest)
+    } else {
+        (RangeOp::Eq, token)
+    };
+
+    parse_semver(rest).map(|version| (op, version))
+}
+
+pub(crate) fn compare_range(version: SemVer, op: RangeOp, bound: SemVer) -> bool {
+    match op {
+        RangeOp::Gt => version > bound,
+        RangeOp::Gte => version >= bound,
+        RangeOp::Lt => version < bound,
+        RangeOp::Lte => version <= bound,
+        RangeOp::Eq => version == bound,
+    }
+}
+
+pub(crate) fn parse_semver(value: &str) -> Option<SemVer> {
+    let value = value.trim();
+    if value.is_empty() {
+        return None;
+    }
+    let core = value.split(['-', '+']).next().unwrap_or(value);
+    let mut parts = core.split('.');
+    let major: u32 = parts.next()?.parse().ok()?;
+    let minor: u32 = parts.next().unwrap_or("0").parse().ok()?;
+    let patch: u32 = parts.next().unwrap_or("0").parse().ok()?;
+    Some(SemVer {
+        major,
+        minor,
+        patch,
+    })
+}
+
+pub(crate) fn resolve_exports_subpath(
+    exports: &serde_json::Value,
+    subpath_key: &str,
+    conditions: &[&str],
+    compiler_version: SemVer,
+) -> Option<String> {
+    match exports {
+        serde_json::Value::String(value) => (subpath_key == ".").then(|| value.clone()),
+        serde_json::Value::Array(list) => {
+            for entry in list {
+                if let Some(resolved) =
+                    resolve_exports_subpath(entry, subpath_key, conditions, compiler_version)
+                {
+                    return Some(resolved);
+                }
+            }
+            None
+        }
+        serde_json::Value::Object(map) => {
+            let has_subpath_keys = map.keys().any(|key| key.starts_with('.'));
+            if has_subpath_keys {
+                if let Some(value) = map.get(subpath_key)
+                    && let Some(target) =
+                        resolve_exports_target(value, conditions, compiler_version)
+                {
+                    return Some(target);
+                }
+
+                let mut best_match: Option<(usize, String, &serde_json::Value)> = None;
+                for (key, value) in map {
+                    let Some(wildcard) = match_exports_subpath(key, subpath_key) else {
+                        continue;
+                    };
+                    let specificity = key.len();
+                    let is_better = match &best_match {
+                        None => true,
+                        Some((best_len, _, _)) => specificity > *best_len,
+                    };
+                    if is_better {
+                        best_match = Some((specificity, wildcard, value));
+                    }
+                }
+
+                if let Some((_, wildcard, value)) = best_match
+                    && let Some(target) =
+                        resolve_exports_target(value, conditions, compiler_version)
+                {
+                    return Some(apply_exports_subpath(&target, &wildcard));
+                }
+
+                None
+            } else if subpath_key == "." {
+                resolve_exports_target(exports, conditions, compiler_version)
+            } else {
+                None
+            }
+        }
+        _ => None,
+    }
+}
+
+pub(crate) fn resolve_exports_target(
+    target: &serde_json::Value,
+    conditions: &[&str],
+    compiler_version: SemVer,
+) -> Option<String> {
+    match target {
+        serde_json::Value::String(value) => Some(value.clone()),
+        serde_json::Value::Array(list) => {
+            for entry in list {
+                if let Some(resolved) = resolve_exports_target(entry, conditions, compiler_version)
+                {
+                    return Some(resolved);
+                }
+            }
+            None
+        }
+        serde_json::Value::Object(map) => {
+            // Process keys in insertion order (Node.js spec). For each key:
+            // 1. Check if it's a plain condition match
+            // 2. Check if it's a versioned condition like "types@>=1"
+            for (key, value) in map {
+                // Check for versioned condition (e.g., "types@>=1")
+                if let Some(at_pos) = key.find('@') {
+                    let base_condition = &key[..at_pos];
+                    let version_range = &key[at_pos + 1..];
+                    if conditions.contains(&base_condition)
+                        && match_types_versions_range(version_range, compiler_version).is_some()
+                        && let Some(resolved) =
+                            resolve_exports_target(value, conditions, compiler_version)
+                    {
+                        return Some(resolved);
+                    }
+                } else if conditions.contains(&key.as_str())
+                    && let Some(resolved) =
+                        resolve_exports_target(value, conditions, compiler_version)
+                {
+                    return Some(resolved);
+                }
+            }
+            None
+        }
+        _ => None,
+    }
+}
+
+pub(crate) fn resolve_exports_target_candidates(
+    target: &serde_json::Value,
+    conditions: &[&str],
+    compiler_version: SemVer,
+) -> Vec<String> {
+    match target {
+        serde_json::Value::String(value) => vec![value.clone()],
+        serde_json::Value::Array(list) => {
+            let mut candidates = Vec::new();
+            for entry in list {
+                candidates.extend(resolve_exports_target_candidates(
+                    entry,
+                    conditions,
+                    compiler_version,
+                ));
+            }
+            candidates
+        }
+        serde_json::Value::Object(map) => {
+            let mut candidates = Vec::new();
+            for (key, value) in map {
+                if let Some(at_pos) = key.find('@') {
+                    let base_condition = &key[..at_pos];
+                    let version_range = &key[at_pos + 1..];
+                    if conditions.contains(&base_condition)
+                        && match_types_versions_range(version_range, compiler_version).is_some()
+                    {
+                        if value.is_null() {
+                            return Vec::new();
+                        }
+                        candidates.extend(resolve_exports_target_candidates(
+                            value,
+                            conditions,
+                            compiler_version,
+                        ));
+                    }
+                } else if conditions.contains(&key.as_str()) {
+                    if value.is_null() {
+                        return Vec::new();
+                    }
+                    candidates.extend(resolve_exports_target_candidates(
+                        value,
+                        conditions,
+                        compiler_version,
+                    ));
+                }
+            }
+            candidates
+        }
+        _ => Vec::new(),
+    }
+}
+
+pub(crate) fn resolve_imports_subpath_candidates(
+    imports: &serde_json::Value,
+    subpath_key: &str,
+    conditions: &[&str],
+    compiler_version: SemVer,
+) -> Vec<String> {
+    let serde_json::Value::Object(map) = imports else {
+        return Vec::new();
+    };
+
+    let has_subpath_keys = map.keys().any(|key| key.starts_with('#'));
+    if !has_subpath_keys {
+        return Vec::new();
+    }
+
+    if let Some(value) = map.get(subpath_key) {
+        return resolve_exports_target_candidates(value, conditions, compiler_version);
+    }
+
+    let mut best_match: Option<(usize, String, &serde_json::Value)> = None;
+    for (key, value) in map {
+        let Some(wildcard) = match_imports_subpath(key, subpath_key) else {
+            continue;
+        };
+        let specificity = key.len();
+        let is_better = match &best_match {
+            None => true,
+            Some((best_len, _, _)) => specificity > *best_len,
+        };
+        if is_better {
+            best_match = Some((specificity, wildcard, value));
+        }
+    }
+
+    if let Some((_, wildcard, value)) = best_match {
+        return resolve_exports_target_candidates(value, conditions, compiler_version)
+            .into_iter()
+            .map(|target| apply_exports_subpath(&target, &wildcard))
+            .collect();
+    }
+
+    Vec::new()
+}
+
+pub(crate) fn match_exports_subpath(pattern: &str, subpath_key: &str) -> Option<String> {
+    let pattern_inner = pattern.strip_prefix("./")?;
+    let subpath = subpath_key.strip_prefix("./")?;
+
+    // A bare "./" exports entry only exposes explicit file-like subpaths such as
+    // "./index.js". It should not manufacture extensionless package subpaths like
+    // "inner/other" that tsc still rejects with TS2307.
+    if pattern == "./" {
+        let has_explicit_extension = Path::new(subpath)
+            .extension()
+            .is_some_and(|ext| !ext.is_empty());
+        return has_explicit_extension.then(|| subpath.to_string());
+    }
+
+    // Handle deprecated trailing-slash directory patterns like "./dir/".
+    if !pattern_inner.is_empty() && pattern_inner.ends_with('/') && !pattern.contains('*') {
+        if let Some(rest) = subpath.strip_prefix(pattern_inner) {
+            return Some(rest.to_string());
+        }
+        return None;
+    }
+
+    if !pattern.contains('*') {
+        return None;
+    }
+
+    let star = pattern_inner.find('*')?;
+    let (prefix, suffix) = pattern_inner.split_at(star);
+    let suffix = &suffix[1..];
+
+    if !subpath.starts_with(prefix) || !subpath.ends_with(suffix) {
+        return None;
+    }
+
+    let start = prefix.len();
+    let end = subpath.len().saturating_sub(suffix.len());
+    if end < start {
+        return None;
+    }
+
+    Some(subpath[start..end].to_string())
+}
+
+pub(crate) fn match_imports_subpath(pattern: &str, subpath_key: &str) -> Option<String> {
+    if !pattern.contains('*') {
+        return None;
+    }
+    let pattern = pattern.strip_prefix('#')?;
+    let subpath = subpath_key.strip_prefix('#')?;
+
+    let star = pattern.find('*')?;
+    let (prefix, suffix) = pattern.split_at(star);
+    let suffix = &suffix[1..];
+
+    if !subpath.starts_with(prefix) || !subpath.ends_with(suffix) {
+        return None;
+    }
+
+    let start = prefix.len();
+    let end = subpath.len().saturating_sub(suffix.len());
+    if end < start {
+        return None;
+    }
+
+    Some(subpath[start..end].to_string())
+}
+
+pub(crate) fn apply_exports_subpath(target: &str, wildcard: &str) -> String {
+    if target.contains('*') {
+        target.replace('*', wildcard)
+    } else if target.ends_with('/') {
+        // Trailing-slash directory pattern: append the matched portion
+        format!("{target}{wildcard}")
+    } else {
+        target.to_string()
+    }
+}
